+++ tests/src/utils/hash32.h
@@ -7,7 +7,7 @@
 // all functions are used to hash 32-bit valued arrays (int32, float)
 
 uint32
-hashArray(const uint32* arr, int nx, int sx);
+hashArray(const uint32* arr, size_t nx, int sx);
 
 uint32
 hash2dStridedBlock(const uint32* arr, int sx, int sy);
