+++ tests/src/utils/hash64.h
@@ -7,7 +7,7 @@
 // all functions are used to hash 64-bit valued arrays (int64, double)
 
 uint64
-hashArray(const uint64* arr, int nx, int sx);
+hashArray(const uint64* arr, size_t nx, int sx);
 
 uint64
 hash2dStridedBlock(const uint64* arr, int sx, int sy);
